backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose
passlib[bcrypt]
bcrypt==4.3.0
python-multipart
google-generativeai
python-dotenv